+++ ft_strncmp.c
@@ -1,9 +1,11 @@
-int ft_strncmp(const char *s1, const char *s2, int n)
+#include "libft.h"
+
+int ft_strncmp(const char *s1, const char *s2, size_t n)
 {
     unsigned char   *str1 = (unsigned char*)s1;
     unsigned char   *str2 = (unsigned char*)s2;
 
-    int i;
+    size_t  i;
 
     i = 0;
     while ((str1[i] || str2[i]) && i < n)
